Move the `RATEd` class to RATE_res.hpp.
