+++ include/CppRateRes.hpp
@@ -53,83 +53,6 @@
 
 #include "BS_thread_pool.hpp"
 
-inline std::vector<double> rate_from_kld(const std::vector<double> &log_kld, const double kld_sum) {
-    std::vector<double> RATE(log_kld.size());
-    double log_kld_sum = std::log(kld_sum);
-#pragma omp parallel for schedule(static)
-    for (size_t i = 0; i < log_kld.size(); ++i) {
-	RATE[i] = std::exp(log_kld[i] - log_kld_sum);
-    }
-    return RATE;
-}
-
-inline std::vector<double> rate_from_log_kld(const std::vector<double> &log_kld) {
-    double max_elem = 0.0;
-    // TODO pragma with custom reduction to find maximum
-    for (size_t i = 0; i < log_kld.size(); ++i) {
-	max_elem = (max_elem > log_kld[i] ? max_elem : log_kld[i]);
-    }
-    double tmp_sum = 0.0;
-#pragma omp parallel for schedule(static) reduction(+:tmp_sum)
-    for (size_t i = 0; i < log_kld.size(); ++i) {
-	tmp_sum += std::exp(log_kld[i] - max_elem);
-    }
-    double log_kld_sum = std::log(tmp_sum) + max_elem;
-
-    std::vector<double> RATE(log_kld.size());
-#pragma omp parallel for schedule(static)
-    for (size_t i = 0; i < log_kld.size(); ++i) {
-	RATE[i] = std::exp(log_kld[i] - log_kld_sum);
-    }
-    return RATE;
-}
-
-inline double rate_delta(const std::vector<double> &RATE) {
-    double Delta = 0.0;
-    size_t num_snps = RATE.size();
-#pragma omp parallel for schedule(static) reduction(+:Delta)
-    for (size_t i = 0; i < num_snps; ++i) {
-	Delta += RATE[i]*(std::log(num_snps) + std::log(RATE[i] + 1e-16));
-    }
-
-    return Delta;
-}
-
-inline double delta_to_ess(const double delta) {
-    return std::exp(std::log(1.0) - std::log(1.0 + delta))*100.0;
-}
-
-struct RATEd {
-public:
-    double ESS;
-    double Delta;
-    std::vector<double> RATE;
-    std::vector<double> KLD;
-
-    RATEd() = default;
-
-    RATEd(double _ESS, double _Delta, std::vector<double> _RATE, std::vector<double> _KLD) {
-	this->ESS = _ESS;
-	this->Delta = _Delta;
-	this->RATE = _RATE;
-	this->KLD = _KLD;
-    }
-
-    RATEd(std::vector<double> _log_KLD) {
-	std::transform(_log_KLD.begin(), _log_KLD.end(), std::back_inserter(this->KLD), static_cast<double(*)(double)>(std::exp));
-	this->RATE = rate_from_log_kld(_log_KLD);
-	this->Delta = rate_delta(RATE);
-	this->ESS = delta_to_ess(Delta);
-    }
-
-    RATEd(std::vector<double> _KLD, std::vector<double> _RATE) {
-	this->KLD = _KLD;
-	this->RATE = _RATE;
-	this->Delta = rate_delta(RATE);
-	this->ESS = delta_to_ess(Delta);
-    }
-};
-
 inline std::vector<double> get_col(const std::vector<double> &flat, const size_t n_rows, const size_t n_cols, const size_t col_id) {
     std::vector<double> res(n_rows);
